+++ source/renderer/app/routes-config.js
@@ -18,7 +18,11 @@ export const ROUTES = {
     TRANSACTIONS: '/wallets/:id/transactions',
     SEND: '/wallets/:id/send',
     RECEIVE: '/wallets/:id/receive',
-    SETTINGS: '/wallets/:id/settings',
+    SETTINGS: {
+      ROOT: '/wallets/:id/settings',
+      GENERAL: '/wallets/:id/settings/general',
+      UTXO: '/wallets/:id/settings/utxo',
+    },
   },
   SETTINGS: {
     ROOT: '/settings',